Create code
In this code, we first define the video file and create a video element to display the video. We also define the subtitles data as an array of objects that contain the start and end times for each subtitle, as well as the text for the subtitle.

Next, we create a function called displaySubtitles that will display the appropriate subtitle based on the current time of the video. This function uses the find method to find the subtitle object that corresponds to the current time of the video, and then updates the text content of a div element with the ID subtitle.

Finally, we add an event listener to the video element that calls the displaySubtitles function whenever the timeupdate event is triggered. We also add the video and subtitle elements to the DOM using the appendChild method.